+++ src/sass/pages/_schoolContent.scss
@@ -1,0 +1,26 @@
+
+@media (max-width: 600px) {
+    .content__container {
+        height: 550px;
+    }
+    .content__title{
+        .content__fecha {
+            display: block;
+            text-align: center;
+            margin: 1rem;
+        }
+        .content__hora {
+            display: block;
+            text-align: center;
+            margin: 1rem;
+
+        }
+        .content__lugar {
+            text-align: center;
+        }
+    }
+    .content__description {
+        box-shadow: 9px 9px 16px rgba(163, 177, 198, 0.6), -9px -9px 16px rgba(255, 255, 255, 0.5);
+    }
+    
+}
